# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI Orchestration API
  description: |
    LLM/ツールを組み合わせたDAGワークフローを設計・実行・監視するAPIです。
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8090/api/v1
    description: Development server
security:
  - bearerAuth: []
...
paths:
...
  /workflows/{id}/runs:
    parameters:
      - $ref: '#/components/parameters/WorkflowId'
    get:
      tags: [Runs]
      summary: 実行一覧取得
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Run'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
    post:
      tags: [Runs]
      summary: ワークフロー実行
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRunRequest'
      responses:
        '201':
          description: 実行開始
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Run'
        '400':
          $ref: '#/components/responses/SchemaValidationError'
  /runs/{runId}:
    parameters:
      - $ref: '#/components/parameters/RunId'
    get:
      tags: [Runs]
      summary: 実行詳細取得
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/RunWithDetails'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    WorkflowId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    RunId:
      name: runId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Run:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        workflow_id:
          type: string
          format: uuid
        workflow_version:
          type: integer
        status:
          type: string
          enum: [pending, running, completed, failed, cancelled]
        run_number:
          type: integer
          description: ワークフロー毎の連番
        input:
          type: object
        output:
          type: object
        error:
          type: string
        triggered_by:
          type: string
          enum: [manual, schedule, webhook, test, internal]
        triggered_by_user:
          type: string
          format: uuid
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    RunWithDetails:
      allOf:
        - $ref: '#/components/schemas/Run'
        - type: object
          properties:
            step_runs:
              type: array
              items:
                $ref: '#/components/schemas/StepRun'
    StepRun:
      type: object
      properties:
        id:
          type: string
          format: uuid
        step_id:
          type: string
          format: uuid
        step_name:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed, skipped]
        attempt:
          type: integer
        input:
          type: object
        output:
          type: object
        error:
          type: string
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        duration_ms:
          type: integer
    CreateRunRequest:
      type: object
      properties:
        input:
          type: object
        triggered_by:
          type: string
          enum: [manual, test, webhook, schedule, internal]
          default: manual
          description: 実行トリガーの種類
        version:
          type: integer
          default: 0
          description: 実行するワークフローバージョン（0=最新）
        mode:
          type: string
          enum: [test, production]
          deprecated: true
          description: 非推奨。triggered_byを使用してください（mode:testはtriggered_by:testにマップ）
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
...
  responses:
...
    SchemaValidationError:
      description: input_schemaに対する入力データ検証エラー
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code:
                    type: string
                    example: SCHEMA_VALIDATION_ERROR
                  message:
                    type: string
                    example: Input validation failed
                  details:
                    type: object
                    properties:
                      errors:
                        type: array
                        items:
                          type: object
                          properties:
                            field:
                              type: string
                              example: email
                            message:
                              type: string
                              example: email is required